refactor: simplify novel slice..
