--- lib/features/novel/domain/domain.dart
@@ -1,2 +1,0 @@
-// Services.
-export 'services/watch_tags.dart';
