+++ lib/features/novel/models/description_info.dart
