+++ lib/features/novel/models/head_info.dart
