--- lib/features/novel/presentation/presentation.dart
@@ -1,22 +1,0 @@
-// Models.
-export 'models/description_info.dart';
-export 'models/head_info.dart';
-
-// Pages.
-export 'pages/novel_page.dart';
-
-// Providers.
-export 'providers/description_info_provider.dart';
-export 'providers/intermediate_provider.dart';
-export 'providers/novel_selection_provider.dart';
-export 'providers/tags_stream_family.dart';
-
-// Widgets.
-export 'widgets/action_bar.dart';
-export 'widgets/chapter_list.dart';
-export 'widgets/description.dart';
-export 'widgets/novel_head.dart';
-export 'widgets/novel_view.dart';
-export 'widgets/partial_view.dart';
-export 'widgets/status.dart';
-export 'widgets/tags.dart';
