+++ lib/features/novel/providers/intermediate_provider.dart
