+++ lib/features/novel/providers/tags_stream_family.dart
@@ -1,6 +1,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/common/common.dart';
-import 'package:nacht/features/novel/domain/services/watch_tags.dart';
+
+import '../services/watch_tags.dart';
 
 final tagsStreamFamily = StreamProvider.autoDispose
     .family<List<MetaEntryData>, int>((ref, id) async* {
