+++ lib/features/novel/views/novel_page.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/common/common.dart';
 import 'package:nacht/nht/nht.dart';
 import 'package:nacht/widgets/widgets.dart';
 
-import '../presentation.dart';
+import '../providers/intermediate_provider.dart';
+import '../widgets/novel_view.dart';
+import '../widgets/partial_view.dart';
 
 class NovelPage extends ConsumerWidget {
   const NovelPage({
