+++ lib/features/novel/widgets/tags.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../presentation.dart';
+import '../providers/tags_stream_family.dart';
 
 class Tags extends ConsumerWidget {
   const Tags({
